search: move dialog to own folder
